requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-jwt-extended==4.3.1
flask-migrate==3.1.0
python-dotenv==0.19.1
psycopg2-binary==2.9.1
argon2-cffi==21.1.0
celery==5.1.2
redis==3.5.3
tenacity==8.0.1
requests==2.26.0
pytest==6.2.5
flask-swagger-ui==3.36.0
flask-limiter==1.4